+++ dev/dev_load_oldRdata.R
@@ -1,16 +1,21 @@
-
+#'
 #' Script to load old .Rdata created with lost packages.
 #' in "IAMXXX" you need to replace XXX with the version of the package.
-
-rm(list = ls()) ; .rs.restartR() # clear everything
-library(devtools)
+#'
+#' Maxime Jaunatre
+#'
+.rs.restartR() # clear everything
+#'
+#' You need to replace "IAM64 with the name of the missing package"
+#'
 tst <- namespace::makeNamespace("IAM64") # create an empty namespace to load S4
-assign("H2G2", "DON'T PANIC", env = tst)   # require interactive run !
+base::assign("H2G2", "DON'T PANIC", env = tst)   # require interactive run !
 base::namespaceExport(tst, ls(tst))
 IAM64::H2G2
 
-# library(IAMtest64)
-load("C:/Users/mjaunatr/Desktop/Docs passation/inputs/inputs_GG/input2016_newStocks_MNZstatic.RData")
-# I can attribute S4 class to another package. This should be impossible to save my sanity.
+load(paste0("C:/Users/mjaunatr/Desktop/Docs passation/inputs/",
+            "inputs_GG/input2016_newStocks_MNZstatic.RData"))
+# I can attribute S4 class to another package.
+# This should be impossible to save my sanity.
 attributes(class(input2016))$package <- "IAM"
 attributes(class(argum2016))$package <- "IAM"