+++ NEWS.md
@@ -1,5 +1,6 @@
 # IAM 2.0.0
 
+* Added a NEWS.md file to track changes to the package.
 
 ## Breaking change